flytter endringsikonene siden de brukes i begge flater
