+++ apps/nav-veileders-flate/src/components/tiltak/endre-deltakelse-modaler/EndreSluttarsakModal.tsx
@@ -3,21 +3,19 @@ import {
   DeferredFetchState,
   DeltakerStatusAarsakType,
   DeltakerStatusType,
+  EndreDeltakelseType,
   useDeferredFetch
 } from 'deltaker-flate-common'
 import { useState } from 'react'
 import { useAppContext } from '../../../AppContext.tsx'
 import { endreDeltakelseSluttarsak } from '../../../api/api.ts'
-import {
-  BESKRIVELSE_ARSAK_ANNET_MAX_TEGN,
-  EndreDeltakelseType
-} from '../../../api/data/endre-deltakelse-request.ts'
+import { BESKRIVELSE_ARSAK_ANNET_MAX_TEGN } from '../../../api/data/endre-deltakelse-request.ts'
 import { PameldingResponse } from '../../../api/data/pamelding.ts'
 import { ErrorPage } from '../../../pages/ErrorPage.tsx'
 import { getDeltakerStatusAarsakTypeText } from '../../../utils/displayText.ts'
 import { getDeltakerStatusAarsakTyperAsList } from '../../../utils/utils.ts'
 import { ModalFooter } from '../../ModalFooter.tsx'
-import { EndringTypeIkon } from '../EndringTypeIkon.tsx'
+import { EndringTypeIkon } from 'deltaker-flate-common'
 
 interface EndreSluttarsakModalProps {
   pamelding: PameldingResponse
