+++ apps/nav-veileders-flate/src/components/tiltak/forslag/BehandleForslagKnapp.tsx
@@ -1,10 +1,13 @@
 import { Button } from '@navikt/ds-react'
 import { usePameldingContext } from '../PameldingContext.tsx'
 import { useRef, useState } from 'react'
-import { EndreDeltakelseType } from '../../../api/data/endre-deltakelse-request.ts'
 import { PameldingResponse } from '../../../api/data/pamelding.ts'
 import { ModalController } from '../endre-deltakelse-modaler/ModalController.tsx'
-import { AktivtForslag, ForslagEndringType } from 'deltaker-flate-common'
+import {
+  AktivtForslag,
+  EndreDeltakelseType,
+  ForslagEndringType
+} from 'deltaker-flate-common'
 import { util } from 'zod'
 import assertNever = util.assertNever
