+++ packages/deltaker-flate-common/index.ts
@@ -3,11 +3,13 @@ export * from './components/DeltakerStatusInfoTekst'
 export * from './components/UtkastHeader'
 export * from './components/HvaDelesMedArrangor'
 export * from './components/FeilregistrertInfo'
-export * from './components/forslag/ForlengDeltakelseForslagDetaljer.tsx'
-export * from './components/forslag/ForslagDetaljer.tsx'
+export * from './components/EndringTypeIkon'
+export * from './components/forslag/ForlengDeltakelseForslagDetaljer'
+export * from './components/forslag/ForslagDetaljer'
 export * from './hooks/useDeferredFetch'
 export * from './hooks/useFetch'
 export * from './model/deltaker'
+export * from './model/endre-deltaker'
 export * from './model/forslag'
 export * from './utils/displayText'
 export * from './utils/constants'
