+++ packages/deltaker-flate-common/model/endre-deltaker.ts
@@ -1,0 +1,12 @@
+export enum EndreDeltakelseType {
+  IKKE_AKTUELL = 'IKKE_AKTUELL',
+  ENDRE_INNHOLD = 'ENDRE_INNHOLD',
+  ENDRE_BAKGRUNNSINFO = 'ENDRE_BAKGRUNNSINFO',
+  ENDRE_SLUTTARSAK = 'ENDRE_SLUTTARSAK',
+  ENDRE_SLUTTDATO = 'ENDRE_SLUTTDATO',
+  ENDRE_OPPSTARTSDATO = 'ENDRE_OPPSTARTSDATO',
+  FORLENG_DELTAKELSE = 'FORLENG_DELTAKELSE',
+  AVSLUTT_DELTAKELSE = 'AVSLUTT_DELTAKELSE',
+  ENDRE_DELTAKELSESMENGDE = 'ENDRE_DELTAKELSESMENGDE',
+  REAKTIVER_DELTAKELSE = 'REAKTIVER_DELTAKELSE'
+}
